requirements.txt
langchain
langchain-community
langchain-text-splitters
chromadb
python-dotenv
pypdf
fastapi
uvicorn
openai
python-multipart
langchain-nvidia-ai-endpoints
aiosqlite
Pillow
langchain-openai
torch
torchvision
langchain-chroma
langchain-huggingface
sentence-transformers
streamlit
langchain-google-genai
